Fix reload-on-touch/change
As part of a fix for #2656, commit 8f1d0e5 introduced a measure to wait
for processes to finish work while reloading. The code to do this
affects other logic flows was well, breaking the ability to reload uwsgi
when a config file changes.

This moves the code around a bit so #2656 stays fixed, but
reload-on-touch/change works again.

Fixes #2681.